Added support for accumulated JPEG payloads buffered on the X80Pro object, to be decoded when the total payload has been recieved.

+++ src/main/java/edu/ius/robotics/robots/codecs/X80ProJPEG.java
@@ -2,9 +2,9 @@ package edu.ius.robotics.robots.codecs;
 
 public class X80ProJPEG
 {
-	public byte[] decode(byte[] input, byte length)
+	public byte[] decode(byte[] input, int length)
 	{
-		byte[] output = new byte[length];
+		byte[] output = new byte[length]; // any coefficient?
 		
 		return output;
 	}
